+++ c2cache/main.py
@@ -1,3 +1,0 @@
-from caterva2.services.server import main
-
-__all__ = ["main"]